fixed some warnings

+++ vistle/insitu/core/transformArray.h
@@ -121,7 +121,7 @@ namespace insitu {
                         dim[c] > 1 ? --dim[c] : dim[c];
                     perCell = true;
                 }
-                if (dim[0] * dim[1] * dim[2] != n) {
+                if ((size_t)dim[0] * (size_t)dim[1] * (size_t)dim[2] != n) {
                     std::cerr << "vtkArray2Vistle: non-matching grid size: [" << dim[0] << "*" << dim[1] << "*" << dim[2] << "] != " << n << std::endl;
                     return nullptr;
                 }
